requirements.txt
fastapi
uvicorn
requests
pydantic
mysql-connector-python
sentence-transformers
faiss-cpu
numpy
python-dotenv
